Fixed a bug in the timeline view where if the user is scrolled to the top of that view, and sends a tweet, vertical space for the tweet cell appears, but no content is drawn. The cell is blank until something forces it to redraw, such as scrolling it off the screen and then back on. It also does not respond to touch events.
It seems that adding the tweet to both the profile view (is this even relevant anymore?) and the timeline view causes this problem. It appears that after the tweet is inserted, drawRect: is never called. I'm not sure why, though one guess I have is that it has something to do with the fact that there's only one shared cell instance for any given tweet, and this single instance was being added to one visible table view cell (the timeline view), and then to another table view cell that's not visible (the profile view). I think adding it to the second invisible view causes the cell to think it doesn't need to be displayed.

This change seems to break no functionality as the profile view is refreshed when it's displayed anyway, and you can't send a tweet from that view anymore even if it weren't (so the addTweet: function would never be called).